// frontend/src/app/api/model/matchDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LocalTime } from './localTime';


export interface MatchDto { 
    matchId?: number;
    fieldId?: number;
    type?: string;
    pubStatus?: string;
    privStatus?: string;
    matchDate?: string;
    startTime?: LocalTime;
    endTime?: LocalTime;
    pricing?: number;
    organiserId?: string;
}