// frontend/src/app/api/model/fieldDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FieldDto { 
    fieldId?: number;
    siteId?: number;
    isIndoor?: boolean;
    isActive?: boolean;
    maintenanceFromDate?: string;
    maintenanceToDate?: string;
}